requirements.txt
pyopengl
pyglm
pygame
addict